Add table support in markdown.
.md files use recommonmark + sphinx-markdown-tables.

Unfortunately, this combo doesn't work in docstrings, which have to be
parsed to rst first. So we use m2r for docstrings.

m2r has much richer markdown syntax than commonmark, including tables.
Unfortunately I couldnt get it to work on md files.

This commit also starts the Attr class. It is left incomplete and is
only included as a test of the table parsing.